Moved entities out of models folder, removed unused directories, updated README
