+++ src/entities/continuousPickup.entity.ts
