+++ src/entities/exceptionTypes.entity.ts
