+++ src/entities/fareAttributes.entity.ts
