+++ src/entities/fareRules.entity.ts
