+++ src/entities/feedInfo.entity.ts
