+++ src/entities/frequencies.entity.ts
