+++ src/entities/locationTypes.entity.ts
