+++ src/entities/pickupDropoffTypes.entity.ts
