+++ src/entities/routeTypes.entity.ts
