+++ src/entities/stopTimes.entity.ts
