+++ src/entities/trips.entity.ts
