+++ src/entities/wheelchairAccessible.entity.ts
