+++ src/geo/geo.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { GeoController } from './geo.controller';
-import { ShapeGeoms } from 'src/models/entities/shapeGeoms.entity';
-import { Stops } from 'src/models/entities/stops.entity';
+import { ShapeGeoms } from 'src/entities/shapeGeoms.entity';
+import { Stops } from 'src/entities/stops.entity';
 import { StopsService } from './stops.service';
 import { ShapesService } from './shapes.service';
