+++ src/geo/stops.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { getManager, Repository } from 'typeorm';
-import { Stops } from 'src/models/entities/stops.entity';
+import { Stops } from 'src/entities/stops.entity';
 import { getCurrentDay } from 'src/util';
 
 @Injectable()
