+++ src/gtfs/gtfs.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { GtfsService } from './gtfs.service';
 import { GtfsController } from './gtfs.controller';
-import { Agency } from 'src/models/entities/agency.entity';
+import { Agency } from 'src/entities/agency.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Agency])],
