--- src/routes/routes.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { RoutesController } from './routes.controller';
-
-describe('RoutesController', () => {
-  let controller: RoutesController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [RoutesController],
-    }).compile();
-
-    controller = module.get<RoutesController>(RoutesController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
