--- src/routes/routes.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { RoutesService } from './routes.service';
-
-describe('RoutesService', () => {
-  let service: RoutesService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [RoutesService],
-    }).compile();
-
-    service = module.get<RoutesService>(RoutesService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
