+++ src/transit/routes/routes.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Routes } from 'src/models/entities/routes.entity';
-import { Trips } from 'src/models/entities/trips.entity';
+import { Routes } from 'src/entities/routes.entity';
+import { Trips } from 'src/entities/trips.entity';
 import { getCurrentDay } from 'src/util';
 import { RouteRawData } from 'src/interfaces/data';
