+++ src/transit/transit.module.ts
@@ -6,9 +6,9 @@ import { RoutesService } from './routes/routes.service';
 import { RoutesController } from './routes/routes.controller';
 import { LocationService } from './location/location.service';
 import { LocationController } from './location/location.controller';
-import { Routes } from 'src/models/entities/routes.entity';
-import { Trips } from 'src/models/entities/trips.entity';
-import { Agency } from 'src/models/entities/agency.entity';
+import { Routes } from 'src/entities/routes.entity';
+import { Trips } from 'src/entities/trips.entity';
+import { Agency } from 'src/entities/agency.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Agency, Routes, Trips])],
